Compute in float64 everywhere, not just where it happened to hold
Images are read as float32 -- that is what the scanner wrote -- but the
arithmetic on them was float32 too in several places, and in one of them
that is not accurate enough to report.

identity.brain_similarity is the case that mattered. X @ X.T sums over
every masked voxel, 5.2M of them on a 1mm T1, and BLAS sgemm has no
pairwise summation to fall back on. Measured on two real brains: the
diagonal came back 1.0106 and 1.0131 instead of 1, and the off-diagonal
similarity -- the number the function exists to produce -- read 0.614737
against a true 0.607518. A 1.2% error, on a quantity compared at three
decimal places. In float64 the diagonal is exactly 1 and _corr, a wholly
separate code path, now agrees to 0.607518098.

The rest were measured too small to have changed a conclusion, and are
fixed because being right by accident is not a property that survives
editing:

  surface/{ribbon,concat}  map_coordinates and the depth-average
                           accumulator (3e-7 relative -- but this is the
                           single interpolation the pipeline is built on)
  sdc/{pepolar,phasediff}  the Hz -> mm field conversion
  glm/acompcor             the SVD was ALREADY float64, but only because
                           legendre_trends returns a float64 design matrix
                           and lstsq promoted. A change there would have
                           silently dropped it to float32 with nothing to
                           notice. Now explicit.

Storage is unchanged: GIfTI and NIfTI outputs stay float32, which is those
formats' convention. _niimath.merge_frames and coreg.niimath._to_nifti stay
float32 as well -- they marshal formats and do no arithmetic.

+++ lightprep/sdc/pepolar.py
@@ -62,7 +62,7 @@ def _acq_row(pe_direction: str, readout_time: float) -> str:
 def _load_4d(path: Path):
     """Load an image as 4D, promoting a 3D volume to a single-frame series."""
     img = nib.load(path)
-    data = img.get_fdata(dtype=np.float32)
+    data = img.get_fdata(dtype=np.float64)
     if data.ndim == 3:
         data = data[..., np.newaxis]
     elif data.ndim != 4:
@@ -252,10 +252,14 @@ def pepolar(
     ref_img = nib.load(target)
     voxel_pe = float(ref_img.header.get_zooms()[_pe_axis_index(axis)])
     field = nib.load(fmap_on_target)
-    hz = field.get_fdata(dtype=np.float32)
+    # float64 for the Hz -> mm conversion; the images are written float32,
+    # which is the NIfTI convention, but the arithmetic is not done in it.
+    hz = field.get_fdata(dtype=np.float64)
 
     fieldmap = out_dir / "fieldmap_hz.nii.gz"
-    nib.Nifti1Image(hz, field.affine, field.header).to_filename(fieldmap)
+    nib.Nifti1Image(
+        hz.astype(np.float32), field.affine, field.header
+    ).to_filename(fieldmap)
 
     displacement = sign * hz * total_readout_time * voxel_pe
     dmap = out_dir / "displacementmap.nii.gz"
